LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/coop-same-origin-allow-popups-document-write-expected.txt


FAIL coop-same-origin-allow-popups-document-write assert_equals: opener == null expected "true" but got "false"
